format and session utils
